Address review feedback and CI failures
