+++ app/src/ai/blocklist/mod.rs
@@ -78,12 +78,12 @@ pub use child_agent_launch::inherit_child_agent_settings;
 pub use child_agent_launch::{
     PreparedLocalOzChildLaunch, apply_child_agent_model_override, prepare_local_oz_child_launch,
 };
+#[cfg(feature = "tui")]
+pub use context_model::PendingAttachmentSummary;
 #[cfg(not(feature = "tui"))]
 pub(crate) use context_model::block_context_from_terminal_model;
 #[cfg(feature = "tui")]
 pub use context_model::block_context_from_terminal_model;
-#[cfg(feature = "tui")]
-pub use context_model::PendingAttachmentSummary;
 pub use context_model::{AttachmentType, BlocklistAIContextEvent, BlocklistAIContextModel};
 pub(crate) use context_model::{PendingAttachment, PendingFile};
 pub use controller::BlocklistAIController;
